==> 1_subset_data.Rmd <==

```{r}
library(tidyverse)
library(tidylog)
load("~/research/gocd2/data/cleaned_selected.rdata")
# 
# diary <- diary %>% 
#   group_by(person, cycle_nr) %>% 
#   mutate(menses_length = round(sum(menstruation, na.rm = T)),
#     RCD_std = case_when(
#     round(menstruation)==1 ~ ((RCD + 24 + menses_length)/menses_length * 5) - 24,
#     RCD > -24 ~ RCD,
#     TRUE ~ RCD/cycle_length*29
#   )
# )
# diary %>% filter(RCD_std < -30) %>% select(RCD_std, RCD, cycle_length, menses_length, menstruation)
# ggplot(diary, aes(RCD, RCD_std, color = cycle_length)) + geom_point(alpha = 0.1)

# only singles
diary <- diary %>% 
  mutate(reasons_for_exclusion = str_c(reasons_for_exclusion,
                                            if_else(hetero_relationship == 1 | relationship_status != 1, "not_single, ", "", "")
                                       )
  )
# no women who live with their parents
diary <- diary %>% 
  mutate(reasons_for_exclusion = str_c(reasons_for_exclusion,
                                            if_else(abode_flat_share == 2, "living_with_parents, ", "", "")
                                       )
  )

all_surveys <- all_surveys %>% 
  mutate(reasons_for_exclusion = str_c(reasons_for_exclusion,
                                            if_else(abode_flat_share == 2, "living_with_parents, ", "", "")
                                       )
  )

diary_social <- diary_social %>% mutate(
  person_is_related_woman = if_else(person_relationship_to_anchor == "biological_relative" & person_sex == 1, 1, 0),  
  person_is_unrelated_woman = if_else(person_relationship_to_anchor != "biological_relative" & person_sex == 1, 1, 0),  
  person_is_related_man_inferred_seen = if_else(person_is_related_man_inferred == 1 & person_seen == 1, 1, 0),
  person_is_related_man_seen = if_else(person_is_related_man == 1 & person_seen == 1, 1, 0),
  person_is_related_man_thoughts = if_else(person_is_related_man == 1 & person_thought_about == 1, 1, 0)
)

diary_persons <- diary %>% left_join(diary_social %>% select(session, created_date, starts_with("person_is_related"), starts_with("person_is_unrelated")) %>% group_by(session, created_date) %>% summarise_all(~sum(., na.rm = T)))

n_distinct(diary$person)
mean(diary_persons$person_is_related_man)
var(diary_persons$person_is_related_man)
mean(diary_persons$person_is_unrelated_man)
var(diary_persons$person_is_unrelated_man)
table(diary_persons$person_is_related_man, exclude = NULL)
table(diary_persons$person_is_unrelated_man, exclude = NULL)


diary_persons$premenstrual_phase_fab = factor(diary_persons$premenstrual_phase_fab)
diary_persons$hormonal_contraception = factor(diary_persons$hormonal_contraception)

saveRDS(diary_persons, "diary_persons.rds")
saveRDS(all_surveys, "persons.rds")
```
