==> 2_analyses_relatedness.Rmd <==
---
pagetitle: "Inbreeding aversion analyses"
author: "Julie Driebe & Ruben Arslan"
date: "28 February 2019"
output: html_document
---

# Inbreeding aversion analyses {.tabset}

Pre-registered hypothesis: 

a. Single women spend less time with male biological relatives during the FP (for paired women, we did not include questions about relative sex at this level of detail for reasons of length, but we would not make a different prediction).

b. All women spend less time with relatives during the FP (this question was asked for all participating women).

## load packages {.tabset}
```{r message=F,warning=F}
source("0_helpers.R")
library(knitr) # um diese dokumente zu generieren
library(formr) # formr paket, kann daten laden
opts_chunk$set(warning = F, message = F, error = T, fig.width = 7, fig.height = 6)
library(lubridate) # mit datum-uhrzeit rechnen
library(pander) # schoen formatierte tabellen
library(ggplot2)
library(tidyr) # daten von lang nach breit und zurueck
library(dplyr) # daten zusammenfassen etc. immer als letztes laden
diary <- readRDS("diary_persons.rds")
library(brms)
```

### contact with related men {.tabset}
```{r}
model2_relatives <- brm(person_is_related_man_seen ~ (premenstrual_phase_fab + menstruation + fertile_fab) * hormonal_contraception + (1 + fertile_fab | person), data = diary %>% filter(reasons_for_exclusion == ""), backend = "cmdstanr", family = poisson(), cores = 4, file ="brms_fits/relatives_prob4")
hypothesis(model2_relatives, "fertile_fab < 0", alpha = 0.01)
hypothesis(model2_relatives, "fertile_fab:hormonal_contraceptionTRUE > 0", alpha = 0.01)

model2_relatives_curve <- brm(person_is_related_man_seen ~ s(RCD_fab,  by = hormonal_contraception, bs = "cc") + (1 | person), data = diary %>% filter(reasons_for_exclusion == ""), backend = "cmdstanr", family = poisson(), cores = 4, file ="brms_fits/relatives_prob_curve")

curve_plot(model2_relatives_curve, "No. of related men seen more than 1h")


model2_relatives_r_curve <- brm(person_is_related_man_seen ~ s(RCD_fab,  by = hormonal_contraception, bs = "cc") + (1 | person), data = diary %>% filter(reasons_for_exclusion == "") %>% group_by(session) %>% filter(sd(person_is_related_man_seen, na.rm = T) > 0), backend = "cmdstanr", family = poisson(), cores = 4, file ="brms_fits/model2_relatives_r_curve")

curve_plot(model2_relatives_r_curve, "No. of related men seen more than 1h\nrequire outcome variance")

sjPlot::tab_model(model2_relatives)
(summary_relatives <- summary(model2_relatives))
summary_relatives$random$person %>% exp()

table(model2_relatives$data$person_is_related_man)

margins <- marginal_effects(model2_relatives)
# margins$fertile_fab
marginal_effects(model2_relatives, conditions = data.frame(fertile_fab = c(0,1)), effects = "fertile_fab")


custom_forest(model2_relatives, pars = "fertile_fab") +
    theme(axis.text.y = element_blank()) + 
    geom_hline(yintercept = 0, linetype = 'dashed', size = 1, color = "black") +
    theme(axis.text.y = element_blank(),
          axis.ticks.y = element_blank(),
          legend.position = c(1,0),
          legend.justification = c(1,0)) +
    facet_null() +
    xlab("Individual slopes") +
    scale_y_continuous("Estimated fertile phase effect + 80% CI", breaks = -2:2)

ggsave("figures/ms_fig_varying.pdf", width = 6, height = 7)
ggsave("figures/ms_fig_varying.png", width = 6, height = 7)
```

### time family (singles) {.tabset}
```{r}
library(brms)
library(bayesplot)
model2_family <- brm(time_family ~ (premenstrual_phase_fab + menstruation + fertile_fab) * hormonal_contraception + (1 + fertile_fab | person), data = diary %>% filter(reasons_for_exclusion == "" | reasons_for_exclusion == "not_single, "), cores = 4, backend = "cmdstanr",file = "brms_fits/model2_family3")
hypothesis(model2_family, "fertile_fab < 0", alpha = 0.01)
hypothesis(model2_family, "fertile_fab:hormonal_contraceptionTRUE > 0", alpha = 0.01)


model2_family_curve <- brm(time_family ~ s(RCD_fab,  by = hormonal_contraception, bs = "cc") + (1 | person), data =  diary %>% filter(reasons_for_exclusion == "" | reasons_for_exclusion == "not_single, "), backend = "cmdstanr", cores = 4, file ="brms_fits/model3_family_curve")

curve_plot(model2_family_curve, "Time spent with family")

sjPlot::tab_model(model2_family, ci.hyphen = ";", show.ci = 0.99, show.icc = FALSE, bpe = "mean")
summary(model2_family)
marginal_effects(model2_family)

custom_forest(model2_family, pars = "fertile_fab") +
    theme(axis.text.y = element_blank()) + 
    geom_hline(yintercept = 0, linetype = 'dashed', size = 1, color = "black") +
    theme(axis.text.y = element_blank(),
          axis.ticks.y = element_blank(),
          legend.position = c(0,1),
          legend.justification = c(0,1)) +
    facet_null() +
    ylab("") +
    scale_y_continuous("Est. fertile phase effect + 80% CI", breaks = -2:2)
```

### unrelated men {.tabset}
```{r}
model2_unrelated_men <- brm(person_is_unrelated_man ~ (premenstrual_phase_fab + menstruation + fertile_fab) * hormonal_contraception + (1 + fertile_fab | person), data =  diary %>% filter(reasons_for_exclusion == ""), backend = "cmdstanr", family = poisson(), cores = 4, file ="brms_fits/unrelated_men")

sjPlot::tab_model(model2_unrelated_men, ci.hyphen = ";", show.ci = 0.99, show.icc = FALSE, bpe = "mean")
summary(model2_unrelated_men)
marginal_effects(model2_unrelated_men)
```

### related women {.tabset}
```{r}
model2_related_women <- brm(person_is_related_woman ~ (premenstrual_phase_fab + menstruation + fertile_fab) * hormonal_contraception + (1 + fertile_fab | person), data = diary, backend = "cmdstanr", family = poisson(), cores = 4, file ="brms_fits/related_woman")

sjPlot::tab_model(model2_related_women, ci.hyphen = ";", show.ci = 0.99, show.icc = FALSE, bpe = "mean")
summary(model2_related_women)
marginal_effects(model2_related_women)
```


### unrelated women {.tabset}
```{r}
model2_unrelated_women <- brm(person_is_unrelated_woman ~ (premenstrual_phase_fab + menstruation + fertile_fab) * hormonal_contraception + (1 + fertile_fab | person), data = diary, backend = "cmdstanr", family = poisson(), cores = 4, file ="brms_fits/unrelated_woman")

sjPlot::tab_model(model2_unrelated_women, ci.hyphen = ";", show.ci = 0.99, show.icc = FALSE, bpe = "mean")
summary(model2_unrelated_women)
marginal_effects(model2_unrelated_women)
```
